feat: Refactor CSS injection to a dedicated function for dynamic style loading
